Do not use new grouping with auxiliary image heap or PGO-based sorting.
